# Scripts/swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TodoList.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:5000/"
    }
  ],
  "paths": {
    "/TodoItem": {
      "post": {
        "tags": [
          "TodoItem"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/TodoItem/{id}": {
      "put": {
        "tags": [
          "TodoItem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TodoItem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/TodoList": {
      "get": {
        "tags": [
          "TodoList"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoListDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoListDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoListDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TodoList"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoListCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoListCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoListCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              }
            }
          }
        }
      }
    },
    "/TodoList/{id}": {
      "get": {
        "tags": [
          "TodoList"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TodoList"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoListUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoListUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TodoListUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TodoList"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TodoItemCreateDto": {
        "required": [
          "title",
          "checked",
          "todoListId"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "checked": {
            "type": "boolean"
          },
          "todoListId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "TodoItemDto": {
        "required": [
          "id",
          "title",
          "checked",
          "todoListId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "checked": {
            "type": "boolean"
          },
          "todoListId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "TodoItemUpdateDto": {
        "required": [
          "title",
          "checked"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "checked": {
            "type": "boolean"
          }
        }
      },
      "TodoListCreateDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "TodoListDto": {
        "required": [
          "id",
          "name",
          "todoItems"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "todoItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodoItemDto"
            }
          }
        }
      },
      "TodoListUpdateDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "TodoItem"
    },
    {
      "name": "TodoList"
    }
  ]
}